running script for chimbuko
