+++ test/test.cfg
@@ -18,11 +18,12 @@ ParseMode = Adios
 # - InputFile: input file name, extension must be matched with `Method`
 # ----------------------------------------------------------------------------
 [Adios]
-Method = BP
+Method = SST
 Parameters = verbose=3
 #InputFile = ./data/nwchem/20180801/tau-metrics.bp
 #InputFile = ./data/shortdemo/tau-metrics.bp
-InputFile = ./data/longdemo/tau-metrics-nwchem.bp
+#InputFile = ./data/longdemo/tau-metrics-nwchem.bp
+InputFile = tau-metrics.bp
 
 # ----------------------------------------------------------------------------
 # Define outlider detection algorithm
@@ -80,7 +81,7 @@ VizMethod = offline
 VizUrl = http://0.0.0.0:5000/events
 #VizUrl = http://10.2.228.192:5000/events
 VizPath =
-OutputDir = ./test_dump
+OutputDir = ./json_dump
 
 # ----------------------------------------------------------------------------
 # Debug (or logging) setting (todo: maybe better to rename?)
@@ -94,7 +95,7 @@ OutputDir = ./test_dump
 # ----------------------------------------------------------------------------
 [Debug]
 StopLoop = -1 
-LogFile = test.log
+LogFile = chimbuko.log
 LogFileMode = a 
 LogLevel = INFO
 Format = %(asctime)s - %(processName)s - %(threadName)s - %(filename)s - %(lineno)d - %(levelname)s - %(message)s
